// types/INonfungiblePositionManager.ts
  export declare namespace INonfungiblePositionManager {
    export type CollectParamsStruct = {
      tokenId: PromiseOrValue<BigNumberish>;
      recipient: PromiseOrValue<string>;
      amount0Max: PromiseOrValue<BigNumberish>;
      amount1Max: PromiseOrValue<BigNumberish>;
    };
  
    export type CollectParamsStructOutput = [
      BigNumber,
      string,
      BigNumber,
      BigNumber
    ] & {
      tokenId: BigNumber;
      recipient: string;
      amount0Max: BigNumber;
      amount1Max: BigNumber;
    };
  
    export type DecreaseLiquidityParamsStruct = {
      tokenId: PromiseOrValue<BigNumberish>;
      liquidity: PromiseOrValue<BigNumberish>;
      amount0Min: PromiseOrValue<BigNumberish>;
      amount1Min: PromiseOrValue<BigNumberish>;
      deadline: PromiseOrValue<BigNumberish>;
    };
  
    export type DecreaseLiquidityParamsStructOutput = [
      BigNumber,
      BigNumber,
      BigNumber,
      BigNumber,
      BigNumber
    ] & {
      tokenId: BigNumber;
      liquidity: BigNumber;
      amount0Min: BigNumber;
      amount1Min: BigNumber;
      deadline: BigNumber;
    };
  
    export type IncreaseLiquidityParamsStruct = {
      tokenId: PromiseOrValue<BigNumberish>;
      amount0Desired: PromiseOrValue<BigNumberish>;
      amount1Desired: PromiseOrValue<BigNumberish>;
      amount0Min: PromiseOrValue<BigNumberish>;
      amount1Min: PromiseOrValue<BigNumberish>;
      deadline: PromiseOrValue<BigNumberish>;
    };
  
    export type IncreaseLiquidityParamsStructOutput = [
      BigNumber,
      BigNumber,
      BigNumber,
      BigNumber,
      BigNumber,
      BigNumber
    ] & {
      tokenId: BigNumber;
      amount0Desired: BigNumber;
      amount1Desired: BigNumber;
      amount0Min: BigNumber;
      amount1Min: BigNumber;
      deadline: BigNumber;
    };
  
    export type MintParamsStruct = {
      token0: PromiseOrValue<string>;
      token1: PromiseOrValue<string>;
      fee: PromiseOrValue<BigNumberish>;
      tickLower: PromiseOrValue<BigNumberish>;
      tickUpper: PromiseOrValue<BigNumberish>;
      amount0Desired: PromiseOrValue<BigNumberish>;
      amount1Desired: PromiseOrValue<BigNumberish>;
      amount0Min: PromiseOrValue<BigNumberish>;
      amount1Min: PromiseOrValue<BigNumberish>;
      recipient: PromiseOrValue<string>;
      deadline: PromiseOrValue<BigNumberish>;
    };
  
    export type MintParamsStructOutput = [
      string,
      string,
      number,
      number,
      number,
      BigNumber,
      BigNumber,
      BigNumber,
      BigNumber,
      string,
      BigNumber
    ] & {
      token0: string;
      token1: string;
      fee: number;
      tickLower: number;
      tickUpper: number;
      amount0Desired: BigNumber;
      amount1Desired: BigNumber;
      amount0Min: BigNumber;
      amount1Min: BigNumber;
      recipient: string;
      deadline: BigNumber;
    };
  }